requirements.txt
boto3>=1.28.0
pandas>=1.5.0
python-dotenv>=1.0.0
gradio>=4.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
yfinance>=0.2.0
tqdm>=4.65.0
sec-parser>=0.50.0
deep-translator>=1.11.4
openai>=1.0.0
spacy>=3.6.0
langdetect>=1.0.9
